+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -1701,6 +1701,7 @@ sub isf_public_interface_schedule_report_storage_role_values {
             extract_field
             latency_counter
             repeat_counter
+            resource_round_robin_pointer
             rule_trigger_payload_source
             rule_trigger_source
             sample_alias